feat: 기존 register 페이지와 연결
